+++ src/main/java/io/probedock/junitee/annotations/UseMock.java
@@ -1,4 +1,4 @@
-package io.probedock.junitee.dependency;
+package io.probedock.junitee.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;